Replace rapidjson by json for modern c++

+++ src/KafkaW/BrokerSettings.h
@@ -16,7 +16,7 @@ public:
   void apply(rd_kafka_conf_t *RdKafkaConfiguration);
   std::string Address;
   int PollTimeoutMS = 100;
-  std::map<std::string, int> ConfigurationIntegers;
+  std::map<std::string, int64_t> ConfigurationIntegers;
   std::map<std::string, std::string> ConfigurationStrings;
 };
 }
